extract icon component
